Move findTopLevelTemplates and related functions to templates package
